magento-engcom/community-portal#35 Create Backend API with the projects data
- update validation for the endpoints
