+++ functions/handlers/project/get/getProjectDetails/handler.ts
@@ -4,6 +4,7 @@ import awsSdk = require('aws-sdk');
 const serverless = require('serverless-http');
 const bodyParser = require('body-parser');
 const app = express();
+const { validator } = require('./../../../../lib/validator');
 const utils = require('./../../../../lib/utils');
 
 const dynamodb = utils.dynamodb;
@@ -12,11 +13,19 @@ const PROJECTS_TABLE = process.env.PROJECTS_TABLE;
 app.use(bodyParser.json({ strict: false }));
 
 app.get('/project/id/:project_id/', (req:express.Request, res:express.Response) => {
+  const data = req.params;
+
+  // validate input
+  const valid = validator.validate('projectIdOnlySchema', data);
+
+  if (!valid) {
+    res.status(400).json({ errors: validator.errors });
+    return;
+  }
+
   const params = {
     TableName: PROJECTS_TABLE,
-    Key: {
-      project_id: req.params.project_id,
-    },
+    Key: data,
   };
 
   dynamodb.get(params, (error: Error, result: any) => {
